# ring: exit early if there aren't enough zones or instances