**What this PR does**:

Fixes a performance regression when there aren't enough instances in the ring to meet the replication factor. This may happen when excluding a zone via `ExcludedZones`.

What used to happen in that case is the `Ring.Get` method iterated over all tokens in the ring before existing. This PR adds a condition to exit early when we know that we can't gather enough instances.

<details>
<summary>Benchmark</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/dskit/ring
                                                         │ before-96c85c2.txt │          after-0719614.txt          │
                                                         │       sec/op       │    sec/op     vs base               │
Ring_Get/with_zone_awareness-10                                  795.6n ± 64%   761.6n ± 23%        ~ (p=0.240 n=6)
Ring_Get/one_excluded_zone-10                                1533633.0n ±  9%   570.3n ± 22%  -99.96% (p=0.002 n=6)
Ring_Get/without_zone_awareness-10                               511.2n ±  0%   522.0n ±  1%   +2.11% (p=0.002 n=6)
Ring_Get/without_zone_awareness,_not_enough_instances-10       21114.0n ±  3%   421.7n ±  0%  -98.00% (p=0.002 n=6)
geomean                                                          10.71µ         556.1n        -94.81%

                                                         │ before-96c85c2.txt │         after-0719614.txt          │
                                                         │        B/op        │    B/op     vs base                │
Ring_Get/with_zone_awareness-10                                  0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Ring_Get/one_excluded_zone-10                                    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Ring_Get/without_zone_awareness-10                               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Ring_Get/without_zone_awareness,_not_enough_instances-10         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                                     ²               +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                         │ before-96c85c2.txt │         after-0719614.txt          │
                                                         │     allocs/op      │ allocs/op   vs base                │
Ring_Get/with_zone_awareness-10                                  0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Ring_Get/one_excluded_zone-10                                    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Ring_Get/without_zone_awareness-10                               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
Ring_Get/without_zone_awareness,_not_enough_instances-10         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                                     ²               +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```


</details>

**Checklist**
- [x] Tests updated
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`